scholar-helper/pyproject.toml
[project]
name = "scholar-helper-app"
version = "0.1.0"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
streamlit = "1.51.0"
st-pages = "1.0.1"
pandas = "2.2.3"
plotly = "5.24.1"
requests = "2.32.3"
urllib3 = "2.2.2"
psycopg2-binary = "2.9.10"
sqlalchemy = "2.0.40"
toml = "0.10.2"
python-dateutil = "2.9.0.post0"
httpx = "0.25.2"
cachetools = "5.3.3"
python-dotenv = "1.0.1"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
